Add  Timer.js file for organisation
